requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.32.0

telethon>=1.36.0
PyGithub>=2.3.0

openai>=1.40.0